refactor: code cleanup
- Separated structs
    - Nested card types into further folder

- Added more code documentation

- General UI cleanup
