+++ src/models/card_types/mod.rs
@@ -1,0 +1,5 @@
+pub mod fill_in_the_blanks;
+pub mod flashcard;
+pub mod multiple_answer;
+pub mod multiple_choice;
+pub mod order;
