+++ src/models/mod.rs
@@ -1,7 +1,5 @@
 pub mod card;
-pub mod fill_in_the_blanks;
-pub mod flashcard;
-pub mod multiple_answer;
-pub mod multiple_choice;
-pub mod order;
+pub mod card_types;
+pub mod cards;
+pub mod choice;
 pub mod stateful_list;
