+++ src/ui.rs
@@ -10,16 +10,21 @@ use tui::{
 use crate::{models::card::Card, AppState};
 
 pub fn ui<B: Backend>(f: &mut Frame<B>, app_state: &mut AppState) {
+    let card_question;
+
     let size = f.size();
 
+    // A helper closure to create blocks
     let create_block = |title: &str| {
         Block::default()
             .borders(Borders::ALL)
             .title(title.to_string())
     };
 
+    // The main canvas
     let chunks = Layout::default()
         .horizontal_margin(2)
+        // Card (90%), spacer (5%), controls (5%)
         .constraints([
             Constraint::Percentage(90),
             Constraint::Percentage(5),
@@ -27,27 +32,26 @@ pub fn ui<B: Backend>(f: &mut Frame<B>, app_state: &mut AppState) {
         ])
         .split(size);
 
-    let inner_card = Layout::default()
+    // The area held for the card
+    let card_layout = Layout::default()
         .margin(size.area() / 800)
+        // Title (30%) and content (70%)
         .constraints([Constraint::Percentage(30), Constraint::Percentage(70)])
         .split(chunks[0]);
 
-    let content_layout = Layout::default()
+    // The area held within each card
+    let inner_card_layout = Layout::default()
         .horizontal_margin(size.area() / 800)
+        // Content (90%) and card footer (10%)
         .constraints([Constraint::Percentage(90), Constraint::Percentage(10)])
-        .split(inner_card[1]);
+        .split(card_layout[1]);
 
+    // Create card footer content
     let incorrect = Paragraph::new(Span::styled(
         app_state.incorrect_answers.to_string(),
         Style::default().fg(Color::Red),
     ))
     .alignment(Alignment::Left);
-
-    let correct = Paragraph::new(Span::styled(
-        app_state.correct_answers.to_string(),
-        Style::default().fg(Color::Green),
-    ))
-    .alignment(Alignment::Right);
 
     let cards = Paragraph::new(format!(
         "{}/{}",
@@ -56,12 +60,15 @@ pub fn ui<B: Backend>(f: &mut Frame<B>, app_state: &mut AppState) {
     ))
     .alignment(Alignment::Center);
 
+    let correct = Paragraph::new(Span::styled(
+        app_state.correct_answers.to_string(),
+        Style::default().fg(Color::Green),
+    ))
+    .alignment(Alignment::Right);
+
     match app_state.cards.selected() {
         Card::FlashCard(card) => {
-            let question = Paragraph::new(card.question.to_string())
-                .block(create_block("Question"))
-                .wrap(Wrap { trim: false })
-                .alignment(Alignment::Center);
+            card_question = card.question.clone();
 
             let answer = Paragraph::new(if card.flipped {
                 card.answer.to_string()
@@ -74,15 +81,11 @@ pub fn ui<B: Backend>(f: &mut Frame<B>, app_state: &mut AppState) {
 
             let controls = Paragraph::new("SPACE: Show cards back").alignment(Alignment::Left);
 
-            f.render_widget(question, inner_card[0]);
-            f.render_widget(answer, inner_card[1]);
+            f.render_widget(answer, card_layout[1]);
             f.render_widget(controls, chunks[2]);
         }
         Card::MultipleChoice(card) => {
-            let question = Paragraph::new(card.question.to_string())
-                .block(create_block("Question"))
-                .wrap(Wrap { trim: false })
-                .alignment(Alignment::Center);
+            card_question = card.question.clone();
 
             let choices: Vec<ListItem> = card
                 .choices
@@ -126,17 +129,14 @@ pub fn ui<B: Backend>(f: &mut Frame<B>, app_state: &mut AppState) {
                 .block(create_block("Choices"))
                 .highlight_symbol("> ");
 
-            let controls = Paragraph::new("SPACE: Select choice, ENTER: Validate answer").alignment(Alignment::Left);
+            let controls = Paragraph::new("SPACE: Select choice, ENTER: Validate answer")
+                .alignment(Alignment::Left);
 
-            f.render_widget(question, inner_card[0]);
-            f.render_stateful_widget(choices_list, inner_card[1], &mut card.choices.state);
+            f.render_stateful_widget(choices_list, card_layout[1], &mut card.choices.state);
             f.render_widget(controls, chunks[2]);
         }
         Card::MultipleAnswer(card) => {
-            let question = Paragraph::new(card.question.to_string())
-                .block(create_block("Question"))
-                .wrap(Wrap { trim: false })
-                .alignment(Alignment::Center);
+            card_question = card.question.clone();
 
             let choices: Vec<ListItem> = card
                 .choices
@@ -158,29 +158,21 @@ pub fn ui<B: Backend>(f: &mut Frame<B>, app_state: &mut AppState) {
             let controls =
                 Paragraph::new("SPACE: Select/unselect choice").alignment(Alignment::Left);
 
-            f.render_widget(question, inner_card[0]);
-            f.render_stateful_widget(choices_list, inner_card[1], &mut card.choices.state);
+            f.render_stateful_widget(choices_list, card_layout[1], &mut card.choices.state);
             f.render_widget(controls, chunks[2]);
         }
         Card::FillInTheBlanks(card) => {
-            let question = Paragraph::new(card.question.to_string())
-                .block(create_block("Question"))
-                .wrap(Wrap { trim: false })
-                .alignment(Alignment::Center);
+            card_question = card.question.clone();
 
             let content = Paragraph::new(card.content.to_string())
                 .block(create_block("Content"))
                 .wrap(Wrap { trim: false })
                 .alignment(Alignment::Center);
 
-            f.render_widget(question, inner_card[0]);
-            f.render_widget(content, inner_card[1]);
+            f.render_widget(content, card_layout[1]);
         }
         Card::Order(card) => {
-            let question = Paragraph::new(card.question.to_string())
-                .block(create_block("Question"))
-                .wrap(Wrap { trim: false })
-                .alignment(Alignment::Center);
+            card_question = card.question.clone();
 
             let choices: Vec<ListItem> = card
                 .shuffled
@@ -217,13 +209,22 @@ pub fn ui<B: Backend>(f: &mut Frame<B>, app_state: &mut AppState) {
             )
             .alignment(Alignment::Left);
 
-            f.render_widget(question, inner_card[0]);
-            f.render_stateful_widget(choices_list, inner_card[1], &mut card.shuffled.state);
+            f.render_stateful_widget(choices_list, card_layout[1], &mut card.shuffled.state);
             f.render_widget(controls, chunks[2]);
         }
     };
 
-    f.render_widget(incorrect, content_layout[1]);
-    f.render_widget(cards, content_layout[1]);
-    f.render_widget(correct, content_layout[1]);
+    // Render card title
+    f.render_widget(
+        Paragraph::new(card_question)
+            .block(create_block("Question"))
+            .wrap(Wrap { trim: false })
+            .alignment(Alignment::Center),
+        card_layout[0],
+    );
+
+    // Render card footer
+    f.render_widget(incorrect, inner_card_layout[1]);
+    f.render_widget(cards, inner_card_layout[1]);
+    f.render_widget(correct, inner_card_layout[1]);
 }
